+++ packages/stream_core_flutter/lib/src/theme/stream_theme_extensions.dart
@@ -14,6 +14,7 @@ import 'components/stream_context_menu_action_theme.dart';
 import 'components/stream_context_menu_theme.dart';
 import 'components/stream_emoji_button_theme.dart';
 import 'components/stream_emoji_chip_theme.dart';
+import 'components/stream_error_badge_theme.dart';
 import 'components/stream_jump_to_unread_button_theme.dart';
 import 'components/stream_list_tile_theme.dart';
 import 'components/stream_media_viewer_theme.dart';
@@ -141,6 +142,9 @@ extension StreamThemeExtension on BuildContext {
 
   /// Returns the [StreamEmojiChipThemeData] from the nearest ancestor.
   StreamEmojiChipThemeData get streamEmojiChipTheme => StreamEmojiChipTheme.of(this);
+
+  /// Returns the [StreamErrorBadgeThemeData] from the nearest ancestor.
+  StreamErrorBadgeThemeData get streamErrorBadgeTheme => StreamErrorBadgeTheme.of(this);
 
   /// Returns the [StreamJumpToUnreadButtonThemeData] from the nearest ancestor.
   StreamJumpToUnreadButtonThemeData get streamJumpToUnreadButtonTheme => StreamJumpToUnreadButtonTheme.of(this);